# openapi.yaml
openapi: 3.1.0
info:
  title: Central de Serviços API
  description: API para gestão de serviços integrada ao Omie
  version: 1.0.5
  contact:
    name: Fabio
    email: fabio@oondemand.com.br
servers:
  - url: https://api.oondemand.com.br/central-servicos
    description: Servidor de Produção
components:
  schemas:
    Pessoa:
      type: object
      properties:
        documento:
          type: string
          description: Documento da pessoa (CPF ou CNPJ)
          pattern: '^\d{11}$|^\d{14}$'
        tipo:
          type: string
          enum: [pf, pj]
        nome:
          type: string
          maxLength: 100
    Valores:
      type: object
      properties:
        aliquota:
          type: number
        valorIss:
          type: number
        valorServicos:
          type: number
    Servico:
      type: object
      properties:
        valores:
          $ref: '#/components/schemas/Valores'
        issRetido:
          type: number
        discriminacao:
          type: string
          maxLength: 500
    DeclaracaoPrestacaoServico:
      type: object
      properties:
        competencia:
          type: string
          description: Competência no formato MM/YYYY
        servico:
          $ref: '#/components/schemas/Servico'
        optanteSimplesNacional:
          type: number
    InfoNfse:
      type: object
      properties:
        numero:
          type: number
        dataEmissao:
          type: string
          format: date
        prestador:
          $ref: '#/components/schemas/Pessoa'
        tomador:
          $ref: '#/components/schemas/Pessoa'
        declaracaoPrestacaoServico:
          $ref: '#/components/schemas/DeclaracaoPrestacaoServico'
    Nfse:
      type: object
      properties:
        arquivoOriginal:
          type: string
          description: Arquivo XML ou PDF original em base64
        infoNfse:
          $ref: '#/components/schemas/InfoNfse'
    Ticket:
      type: object
      properties:
        titulo:
          type: string
        observacao:
          type: string
        etapa:
          type: string
        status:
          type: string
        nfse:
          type: string
          description: ID da NFSe associada
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /nfse:
    get:
      summary: Listar NFSe
      operationId: listNfse
      responses:
        '200':
          description: Lista de NFSe
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Nfse'
    post:
      summary: Criar NFSe
      operationId: createNfse
      requestBody:
        description: Dados da NFSe
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nfse'
      responses:
        '201':
          description: NFSe criada com sucesso
  /nfse/{id}:
    get:
      summary: Obter NFSe por ID
      operationId: getNfseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: NFSe encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Nfse'
        '404':
          description: NFSe não encontrada

    put:
      summary: Atualizar NFSe
      operationId: updateNfse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Dados atualizados da NFSe
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nfse'
      responses:
        '200':
          description: NFSe atualizada com sucesso
        '404':
          description: NFSe não encontrada

    delete:
      summary: Excluir NFSe
      operationId: deleteNfse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NFSe excluída com sucesso
        '404':
          description: NFSe não encontrada
  /tickets:
    get:
      summary: Listar todos os tickets
      operationId: listTickets
      responses:
        '200':
          description: Lista de tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
    post:
      summary: Criar novo ticket
      operationId: createTicket
      requestBody:
        description: Dados do ticket
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
      responses:
        '201':
          description: Ticket criado com sucesso
  /tickets/{id}:
    get:
      summary: Obter ticket por ID
      operationId: getTicketById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket não encontrado

    put:
      summary: Atualizar ticket
      operationId: updateTicket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Dados atualizados do ticket
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
      responses:
        '200':
          description: Ticket atualizado com sucesso
        '404':
          description: Ticket não encontrado

    delete:
      summary: Excluir ticket
      operationId: deleteTicket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Ticket excluído com sucesso
        '404':
          description: Ticket não encontrado
...